# api.yaml
openapi: 3.1.0
info:
  title: Mini-Scale E-Commerce API
  summary: Canonical contract for a small B2C marketplace (up to ~30 users).
  version: 0.1.0
  description: |
    API-first specification for a multi-vendor e-commerce website.
    - JSON only; UTF-8; all timestamps are RFC 3339 (`YYYY-MM-DDTHH:MM:SSZ`).
    - Authorization via Bearer JWT (access + refresh).
    - Idempotency on payment/checkout critical calls via `Idempotency-Key` header.
    - Pagination uses page/size + metadata; filtering and sorting are standardized parameters.
  termsOfService: https://example.com/terms
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

jsonSchemaDialect: https://spec.openapis.org/oas/3.1/dialect/base

servers:
  - url: https://{env}.example.com/api/v1
    description: Environment server
    variables:
      env:
        default: staging
        enum: [local, staging, prod]

tags:
  - name: Health
    description: Liveness/readiness probes.
  - name: Auth
    description: Registration, login, token refresh, email verification, password flows.

security:
  - bearerAuth: []

# ======================================================================
paths:
  /secure-health:
    get:
      tags: [Health]
      summary: Liveness probe
      operationId: getHealth
      responses:
        '200':
          description: Service is up
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, enum: [ok] }
                  time: { type: string, format: date-time }
              examples:
                ok:
                  value: { status: ok, time: "2025-01-01T00:00:00Z" }

  # -------------------------
  # Auth
  # -------------------------
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a customer account
      operationId: register
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            examples:
              default:
                value:
                  email: user@example.com
                  password: Str0ngP@ss!
                  name: "Jane Doe"
      responses:
        '201':
          description: Registered; email verification may be required
          headers:
            X-Request-Id: { $ref: '#/components/headers/X-Request-Id' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthenticatedResponse' }
        '400': { $ref: '#/components/responses/Error400' }

  /auth/login:
    post:
      tags: [Auth]
      summary: Login with email/password
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthenticatedResponse' }
        '401': { $ref: '#/components/responses/Error401' }

  /auth/refresh:
    post:
      tags: [Auth]
      summary: Rotate tokens using refresh token
      operationId: refresh
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RefreshTokenRequest' }
      responses:
        '200':
          description: New access and refresh tokens
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthenticatedResponse' }
        '401': { $ref: '#/components/responses/Error401' }

  /auth/logout:
    post:
      tags: [Auth]
      summary: Invalidate current access/refresh tokens
      operationId: logout
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RefreshTokenRequest' }
      responses:
        '200':
          description: Logged out
        '401': { $ref: '#/components/responses/Error401' }

  /me/password:
    put:
      tags: [Auth]
      summary: Change password
      operationId: changePassword
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PasswordChangeRequest' }
      responses:
        '204':
          description: Password changed
        '400': { $ref: '#/components/responses/Error400' }

# ======================================================================
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  headers:
    X-Request-Id:
      description: Correlation identifier for tracing requests.
      schema: { type: string }

  responses:
    Error400:
      description: Bad Request
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
    Error401:
      description: Unauthorized
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
    Error403:
      description: Forbidden
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
    Error404:
      description: Not Found
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }

  schemas:

    # ----------------- Core primitives -----------------
    ErrorResponse:
      type: object
      required: [httpStatus, errors]
      properties:
        httpStatus: { type: integer, minimum: 100, maximum: 600 }
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ErrorItem'

    ErrorItem:
      type: object
      required: [message, exception]
      properties:
        message: { type: string }
        field: { type: string }
        exception: { type: string }
        file: { type: string }
        line: { type: integer }
        trace:
          type: array
          items:
            type: object
            additionalProperties: true

    # ----------------- Users & Auth -----------------
    User:
      type: object
      required: [id, email, firstName, lastName, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        emailVerifiedAt: { type: string, format: date-time }
        password:
          type: string
          minLength: 8
          pattern: '^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)(?=.*[\W_]).+$'
          description: Must contain at least 1 uppercase, 1 lowercase, 1 number, and 1 special char.
        passwordUpdatedAt: { type: string, format: date-time }
        firstName: { type: string }
        lastName: { type: string }
        phone: { type: string }
        roles:
          type: array
          items: { type: string, enum: [role_customer, role_vendor, role_admin] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }

    RegisterRequest:
      type: object
      required: [email, password, firstName, lastName]
      properties:
        email:
          type: string
          format: email
          description: Must be unique in the system.
        password:
          type: string
          minLength: 8
          pattern: '^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)(?=.*[\W_]).+$'
          description: Must contain at least 1 uppercase, 1 lowercase, 1 number, and 1 special char.
        firstName: { type: string, minLength: 1 }
        lastName: { type: string, minLength: 1 }

    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }

    RefreshTokenRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken: { type: string }

    PasswordChangeRequest:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword: { type: string }
        newPassword:
          type: string
          minLength: 8
          pattern: '^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)(?=.*[\W_]).+$'
          description: Must contain at least 1 uppercase, 1 lowercase, 1 number, and 1 special char.

    AuthenticatedResponse:
      type: object
      required: [accessToken, refreshToken]
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }